HW3/Part 1/Server/requirements.txt
Flask
torch
pandas
opencv-python
Pillow
Werkzeug
requests